Rename config to wireguard_config mod
